refactor(cb2-5539): fix unit tests
