+++ src/app/features/test-records/amend/views/test-record/test-record.component.spec.ts
@@ -22,8 +22,8 @@ import { of, ReplaySubject } from 'rxjs';
 import { DynamicFormsModule } from '../../../../../forms/dynamic-forms.module';
 import { BaseTestRecordComponent } from '../../../components/base-test-record/base-test-record.component';
 import { ResultOfTestComponent } from '../../../components/result-of-test/result-of-test.component';
+import { VehicleHeaderComponent } from '../../../components/vehicle-header/vehicle-header.component';
 import { TestAmendmentHistoryComponent } from '../../components/test-amendment-history/test-amendment-history.component';
-import { VehicleHeaderComponent } from '../../components/vehicle-header/vehicle-header.component';
 import { TestRecordComponent } from './test-record.component';
 
 describe('TestRecordComponent', () => {
