+++ src/app/features/test-records/components/base-test-record/base-test-record.component.spec.ts
@@ -12,16 +12,15 @@ import { resultOfTestEnum } from '@models/test-types/test-type.model';
 import { TechnicalRecordService } from '@services/technical-record/technical-record.service';
 import { TestTypesService } from '@services/test-types/test-types.service';
 import { SharedModule } from '@shared/shared.module';
-import { GetTestResultsService, UpdateTestResultsService } from '@api/test-results';
+import { GetTestResultsService, UpdateTestResultsService, DefaultService as CreateTestResultsService } from '@api/test-results';
 import { VehicleHeaderComponent } from '../vehicle-header/vehicle-header.component';
 import { UserService } from '@services/user-service/user-service';
-import {  of } from 'rxjs';
+import { of } from 'rxjs';
 import { Roles } from '@models/roles.enum';
 
 describe('BaseTestRecordComponent', () => {
   let component: BaseTestRecordComponent;
   let fixture: ComponentFixture<BaseTestRecordComponent>;
-
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
@@ -34,6 +33,7 @@ describe('BaseTestRecordComponent', () => {
         TechnicalRecordService,
         UpdateTestResultsService,
         GetTestResultsService,
+        CreateTestResultsService,
         {
           provide: UserService,
           useValue: {
