+++ src/app/features/test-records/components/vehicle-header/vehicle-header.component.spec.ts
@@ -5,6 +5,7 @@ import { provideMockStore } from '@ngrx/store/testing';
 import { ResultOfTestService } from '@services/result-of-test/result-of-test.service';
 import { SharedModule } from '@shared/shared.module';
 import { initialAppState } from '@store/.';
+import { of } from 'rxjs';
 import { ResultOfTestComponent } from '../result-of-test/result-of-test.component';
 import { VehicleHeaderComponent } from './vehicle-header.component';
 
@@ -28,5 +29,17 @@ describe('VehicleHeaderComponent', () => {
 
   it('should create', () => {
     expect(component).toBeTruthy();
+  });
+
+  it('should combine the odometer reading', () => {
+    expect(component.combinedOdometerReading(1234, 'kilometres')).toEqual('1234 km');
+  });
+
+  it('should display the unit if the reading is undefined', () => {
+    expect(component.combinedOdometerReading(undefined, 'kilometres')).toEqual(' km');
+  });
+
+  it('should display the reading if the unit is undefined', () => {
+    expect(component.combinedOdometerReading(1234, undefined)).toEqual('1234 ');
   });
 });
